Framework/Media
Brought in some of the windows-specific changes because I believe some
of the underlying changes will require bringing these back before we
can have an operational editor.

Also continuing with normal processing...

I'm currently at a loss as to what to do with the "Destructive"
commands/addons as they were originally written specifically with
Windows in mind.  I believe we will destroy them entirely and
bring in Haiku-specific capabilties later.

Once things are building and running, we will need to flatten out
unnecessary inheritence heirarchies which are related purely to
portability - which is a HUGE amount of Titan's overhead.

This may entail an entirely new GUI, however the framework here for
the GUI is actually fairly decent and mates up quite well with
Haiku API, so it may be something that can be addressed with
automated utilities...