Continued BObjectList conversion.
Had to run some simulated loops to see what was the best route,
using BObjectList is, indeed, the best route for the long-term
even though performance pales in comparison to c-arrays.

The STL crap is going to be ripped out as well in preference to
BList, BObjectList, and custom containers with a compatible
interface as needed.